Move QuickLink to component folder
